tweet-fetcher/src/requirements.txt
pika==1.2.1
tweepy==4.10.0
python-dateutil